+++ src/components/interfaces.ts
@@ -17,11 +17,3 @@ export type Width = "auto" | "half" | "full";
 // used to help track of event payloads to remove at 1.0.0 – see https://github.com/Esri/calcite-components/issues/3781
 /* Note : should be removed before `1.0 */
 export type DeprecatedEventPayload = any;
-
-export enum KindIcons {
-  brand = "lightbulb",
-  danger = "exclamationMarkTriangle",
-  info = "information",
-  success = "checkCircle",
-  warning = "exclamationMarkTriangle"
-}
