+++ src/components/resources.ts
@@ -1,0 +1,7 @@
+export enum KindIcons {
+  brand = "lightbulb",
+  danger = "exclamationMarkTriangle",
+  info = "information",
+  success = "checkCircle",
+  warning = "exclamationMarkTriangle"
+}
